Move store() and store_raw() into a new module and file.
store_raw() is huge and clutters up the root module, so move it. store()
is tightly coupled, so move that too.

No functional change.
